updates the S3Store so it can be used as a DiskStore without chaining
The ChainedStore was commonly used with a DiskStore and S3Store. With
this commit that is no longer needed. If you still do use them together
in a ChainedStore you will need to set always_check_s3 to True for the
S3Store. In general the ChainedStore is a lot less useful now. The only
usecase I can think of maybe using it is with multiple DiskStores. That
said, if you are doing that you are probably using ChainedRepos and so
you wouldn't need a ChainedStore as well. In short, the ChainedStore may
be deprecated soon. This commit is a stopgap fix until we do.